django_blog/requirements.txt
Django
django-taggit
